Cleans up EntityCoordinates, split coordinates into their own files.
